refactor: reorganize claim dto filesystem
