+++ src/claims/dto/store/base-claim.dto.ts
@@ -8,10 +8,9 @@ import {
 import { IsDID } from 'src/validateutils';
 import { Type } from 'class-transformer';
 import { did_t, claimStatus_t, careerType_t, career_t } from 'did-core';
-import { ClaimContentDto } from './claim-content.dto';
+import { ClaimContentDto } from '../nested/claim-content.dto';
 
 export class BaseClaimDto {
-
     // -------------------------
     // Required fields
     // -------------------------
