+++ src/claims/dto/store/create-claim.dto.ts
@@ -2,8 +2,13 @@ import {
     IsInt,
     IsIn,
     IsEmpty,
+    IsString,
+    IsHash,
 } from 'class-validator'
 import {
+    career_t,
+    CLAIM_STATUS_ACCEPTED_LITERAL,
+    CAREER_TYPE_IPFS_HASH_LITERAL,
     CAREER_TYPE_VC_LITERAL,
 } from 'did-core';
 import Const from 'src/config/const.config';
@@ -26,4 +31,21 @@ export class CreateClaimDto extends BaseClaimDto implements BaseClaimInterface {
 
     @IsEmpty()
     readonly career: undefined;
+}
+
+export class CreateCareerDto extends BaseClaimDto implements BaseClaimInterface {
+    // -------------------------
+    // Implemented fields
+    // -------------------------
+    @IsInt()
+    @IsIn([Const.CLAIM_STATUS_ACCEPTED])
+    readonly status: CLAIM_STATUS_ACCEPTED_LITERAL;
+
+    @IsInt()
+    @IsIn([Const.CAREER_TYPE_IPFS_HASH])
+    readonly careerType: CAREER_TYPE_IPFS_HASH_LITERAL;
+
+    @IsString()
+    @IsHash("sha256")
+    readonly career: career_t;
 }
